ocr-translator/backend/requirements.txt
Flask
pytesseract
pillow
pdf2image
requests
reportlab
werkzeug
